Fix bugs

+++ src/amogel/model/graph_classification.py
@@ -281,7 +281,7 @@ class BinaryLearning(pl.LightningModule):
         
         output = self.forward(x1 , edge_index1 , edge_attr1 , x2 , edge_index2 , edge_attr2 , x3 , edge_index3 , edge_attr3 , batch1_idx , batch2_idx , batch3_idx)
         
-        loss = self.loss(output.squeeze(dim=-1) , y1.squeeze(dim = -1))
+        loss = self.loss(output.squeeze(dim=-1) , y1.squeeze(dim = -1).float())
         acc = self.acc(torch.nn.functional.sigmoid(output).squeeze(dim=-1) , y1.squeeze(dim=-1))
 
         self.log("train_loss" , loss , on_epoch=True , on_step=False , prog_bar=True , batch_size=batch1_idx.shape[0])
